+++ frontend/pages/hosts/details/cards/Vitals/Vitals.tsx
