+++ frontend/pages/hosts/details/cards/Vitals/_styles.scss
