+++ frontend/pages/hosts/details/cards/Vitals/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./Vitals";
